requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.14.0
python-dotenv>=1.0.0
requests>=2.28.0
pydantic>=2.0.0
google-generativeai>=0.3.0
openai>=1.0.0
SQLAlchemy>=2.0.0 